Hawk and Dove Simulation
This is a simulation of the popular Game Theory and Evolutionary Stable Strategy game known as the "The Hawk and Dove Game". As with any other game model, our central question is whether to not DOVE and Hawk can co-exist and if so at what frequency. In this game between two opponents, an individual at each encounter has the alternative to choose whether to play the behaviour of a Hawk or a Dove.

Here is a description of the two alternative behaviours.

Hawk:

    Very aggressive, always fights for some resource.
    Fights between hawks are brutal affairs with the loser being the one who first sustains injury.
    The winner takes sole possession of the resource.

Dove:

    Never fights for a resource. It displays in any conflict and if it is attacked it immediately withdraws before it gets injured.
    In any conflict situation between Dove and Hawk, dove will always lose the resource to a hawk.
    It never gets hurt (never sustains a decrease in fitness) when conf